+++ app/pages/NetworkPicker/index.js
@@ -6,10 +6,6 @@ import connect from 'react-redux/es/connect/connect';
 import * as nodeActions from '../../ducks/node';
 import './network-picker.scss';
 
-const GENESIS_TIME = 1580749200000;
-
-const preGenesis = Date.now() < GENESIS_TIME;
-
 const dummyItems = [
   {
     label: '--',
@@ -20,11 +16,9 @@ const dummyItems = [
 const networks = [
   {
     label: 'Mainnet',
-    disabled: preGenesis
   },
   {
     label: 'Testnet',
-    disabled: preGenesis
   },
   {
     label: 'Simnet',